pyproject.toml
[project]
name = "cambrian-stack"
version = "0.1.0"
description = "Rapid transformer experiments - extensible research codebase"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "sumukshashidhar", email = "sumuks2@illinois.edu" }
]
dependencies = [
    # Data & Tokenization
    "datasets>=4.4.1",
    "transformers>=4.35.0",
    
    # Config & Logging  
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "loguru>=0.7.3",
    "wandb>=0.23.0",
    
    # Utilities
    "rich>=13.0.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "torch>=2.0.0",
    "accelerate>=0.25.0",
]
ml = [
    "torch>=2.0.0",
    "accelerate>=0.25.0",
]
docs = [
    "sphinx>=7.2.0",
    "furo>=2024.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cambrian_stack"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU (deselect with '-m \"not gpu\"')",
]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]
